rename src/renderer/pages
