+++ src/renderer/pages/terms.vue
