+++ test/unit/renderer/pages/vpn.spec.js
@@ -18,7 +18,7 @@
 // @flow
 import { createLocalVue, mount } from '@vue/test-utils'
 import { beforeEach, describe, it, expect } from '../../../helpers/dependencies'
-import Vpn from '../../../../src/renderer/pages/Vpn'
+import Vpn from '../../../../src/renderer/pages/vpn'
 import Vuex, { Store } from 'vuex'
 import DIContainer from '../../../../src/app/di/vue-container'
 import RendererCommunication from '../../../../src/app/communication/renderer-communication'
